code rewrite:
- fixed the alive.txt updating twice and not being restored to zero.
- seperated some classes to increase visibility and seperation in order to support new games in the future.
- increased processing speed by replacing PIL with CV2
